triton_framework/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triton-framework"
version = "0.1.0"
description = "A Triton-based training & inference framework scaffold"
authors = [{ name = "Your Name" }]
requires-python = ">=3.8"
dependencies = [
  "torch>=2.1.0",
  "triton>=2.0.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "tqdm>=4.65",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"triton_framework" = ["configs/*.yaml"]
